api/requirements.txt
aniso8601==7.0.0
click==7.1.2
Flask==1.1.2
Flask-GraphQL==2.0.1
Flask-SQLAlchemy==2.4.3
graphene==2.1.8
graphene-sqlalchemy==2.2.2
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
inflect==4.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
promise==2.3
psycopg2==2.8.5
python-dotenv==0.13.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlacodegen==2.1.0
SQLAlchemy==1.3.17
Werkzeug==1.0.1
